+++ src/prelude.rs
@@ -9,6 +9,7 @@ pub use crate::core::value::{PointOfTheCompass, Value};
 pub use crate::errors::InternalError;
 pub use chrono::prelude::*;
 pub use chrono::Duration;
+pub use chrono::{DateTime, Utc};
 pub use crossbeam::channel::{Receiver, Sender};
 pub use crossbeam::thread::Scope;
 pub use log::{debug, error, info, log, warn, Level as LogLevel};